+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 - Scope pointer events on checkboxes to the miller-columns element (PR #14)
 - Trigger remove-topic event on MillerColumnsSelectedElement (PR #15)
-- Trigger click-topic event on MillerColumnsElement (PR #16)
+- Trigger click event on checkboxes when selecting an item in the MillerColumnsElement (PR #16)
 
 ## 0.1.0